# services/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/me": {
      "get": {
        "operationId": "AuthenticationController.me",
        "parameters": [
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "description": "Optional single bounded UUID; invalid or repeated values fail closed.",
            "schema": { "format": "uuid", "maxLength": 128, "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CurrentSessionDto" } }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "400": {
            "description": "The request was malformed or failed closed validation.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "500": {
            "description": "An unexpected failure was safely mapped.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          }
        },
        "summary": "Read the redacted authenticated session identity",
        "tags": ["auth"]
      }
    },
    "/v1/auth/sign-in": {
      "post": {
        "operationId": "AuthenticationController.signIn",
        "parameters": [
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "description": "Optional single bounded UUID; invalid or repeated values fail closed.",
            "schema": { "format": "uuid", "maxLength": 128, "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SignInDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthSessionDto" } }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "400": {
            "description": "The request was malformed or failed closed validation.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "401": {
            "description": "Credentials were rejected.",
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "500": {
            "description": "An unexpected failure was safely mapped.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "503": {
            "description": "Authentication provider is unavailable.",
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          }
        },
        "summary": "Sign in and issue a short-lived session",
        "tags": ["auth"]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthenticationController.refresh",
        "parameters": [
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "description": "Optional single bounded UUID; invalid or repeated values fail closed.",
            "schema": { "format": "uuid", "maxLength": 128, "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SessionRefreshDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionRefreshResponseDto" }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "400": {
            "description": "The request was malformed or failed closed validation.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "401": {
            "description": "The refresh session was rejected.",
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "500": {
            "description": "An unexpected failure was safely mapped.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "503": {
            "description": "Session persistence is unavailable.",
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          }
        },
        "summary": "Rotate a short-lived session",
        "tags": ["auth"]
      }
    },
    "/v1/auth/sign-out": {
      "post": {
        "operationId": "AuthenticationController.signOut",
        "parameters": [
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "description": "Optional single bounded UUID; invalid or repeated values fail closed.",
            "schema": { "format": "uuid", "maxLength": 128, "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SessionSignOutDto" } }
          }
        },
        "responses": {
          "400": {
            "description": "The request was malformed or failed closed validation.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "401": {
            "description": "The session could not be authenticated.",
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "500": {
            "description": "An unexpected failure was safely mapped.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          },
          "503": {
            "description": "Session persistence is unavailable.",
            "headers": {
              "X-Correlation-Id": {
                "description": "Stable UUID that correlates related requests and errors.",
                "schema": { "format": "uuid", "type": "string" }
              },
              "X-Request-Id": {
                "description": "Unique UUID generated for this HTTP request.",
                "schema": { "format": "uuid", "type": "string" }
              }
            }
          }
        },
        "summary": "Revoke a session and clear browser credentials",
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "DataBreeze Control-Plane API",
    "description": "Versioned public system contract and explicitly operational health routes.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CurrentSessionDto": {
        "type": "object",
        "properties": {
          "userId": { "type": "string", "format": "uuid" },
          "organizationId": { "type": "string", "format": "uuid" },
          "workspaceId": { "type": "string", "format": "uuid" },
          "authorizationEpoch": { "type": "number", "minimum": 1 },
          "mfaRequired": { "type": "boolean" }
        },
        "required": ["userId", "organizationId", "authorizationEpoch", "mfaRequired"]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "ngu***@example.com", "maxLength": 254 },
          "password": { "type": "string", "minLength": 12, "maxLength": 128, "writeOnly": true },
          "clientPlatform": { "type": "string", "enum": ["android", "desktop", "web"] }
        },
        "required": ["email", "password", "clientPlatform"]
      },
      "AuthSessionDto": {
        "type": "object",
        "properties": {
          "sessionId": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "organizationId": { "type": "string", "format": "uuid" },
          "workspaceId": { "type": "string", "format": "uuid" },
          "accessToken": { "type": "string", "minLength": 1, "maxLength": 4096 },
          "refreshToken": { "type": "string", "minLength": 1, "maxLength": 4096 },
          "accessExpiresAt": { "type": "string", "format": "date-time" },
          "securityEpoch": { "type": "number", "minimum": 1 },
          "mfaRequired": { "type": "boolean" }
        },
        "required": [
          "sessionId",
          "userId",
          "organizationId",
          "workspaceId",
          "accessToken",
          "accessExpiresAt",
          "securityEpoch",
          "mfaRequired"
        ]
      },
      "SessionRefreshDto": {
        "type": "object",
        "properties": {
          "clientPlatform": { "type": "string", "enum": ["android", "desktop", "web"] },
          "refreshToken": { "type": "string", "minLength": 1, "maxLength": 4096, "writeOnly": true }
        },
        "required": ["clientPlatform"]
      },
      "SessionRefreshResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": { "type": "string", "format": "uuid" },
          "accessToken": { "type": "string", "minLength": 1, "maxLength": 4096 },
          "refreshToken": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096,
            "writeOnly": true
          },
          "accessExpiresAt": { "type": "string", "format": "date-time" }
        },
        "required": ["sessionId", "accessToken", "accessExpiresAt"]
      },
      "SessionSignOutDto": {
        "type": "object",
        "properties": {
          "clientPlatform": { "type": "string", "enum": ["android", "desktop", "web"] },
          "sessionId": { "type": "string", "format": "uuid" }
        },
        "required": ["clientPlatform", "sessionId"]
      },
...
      "Identifier": {
        "title": "Stable UUID Identifier",
        "description": "An opaque stable UUID identifier.",
        "type": "string",
        "format": "uuid"
      },
      "Revision": {
        "title": "Entity Revision",
        "description": "A positive, monotonically increasing entity revision.",
        "type": "integer",
        "minimum": 1
      },
      "UtcTimestamp": {
        "title": "UTC Timestamp",
        "description": "An RFC 3339 date-time normalized to UTC and terminated by uppercase Z.",
        "type": "string",
        "format": "date-time",
        "pattern": "Z$"
      },
      "ProblemDetails": {
        "title": "Problem Details",
        "description": "A closed RFC 7807-compatible problem document with DataBreeze public error extensions.",
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "status", "code", "correlationId", "retryable"],
        "anyOf": [
          { "properties": { "titleKey": true }, "required": ["titleKey"] },
          { "properties": { "messageKey": true }, "required": ["messageKey"] }
        ],
        "properties": {
          "type": { "type": "string", "format": "uri-reference" },
          "title": { "type": "string", "minLength": 1 },
          "titleKey": { "type": "string", "minLength": 1, "maxLength": 255 },
          "status": { "type": "integer", "minimum": 100, "maximum": 599 },
          "detail": { "type": "string" },
          "instance": { "type": "string", "format": "uri-reference" },
          "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_]{0,127}$" },
          "correlationId": { "$ref": "#/components/schemas/Identifier" },
          "retryable": { "type": "boolean" },
          "messageKey": { "type": "string", "minLength": 1, "maxLength": 255 },
          "fieldErrors": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["field", "code"],
              "properties": {
                "field": { "type": "string", "minLength": 1, "maxLength": 255 },
                "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_]{0,127}$" }
              }
            }
          },
          "retryAfterSeconds": { "type": "integer", "minimum": 0 },
          "currentRevision": { "$ref": "#/components/schemas/Revision" },
          "remediationAction": { "type": "string", "minLength": 1, "maxLength": 255 },
          "rateLimit": {
            "type": "object",
            "additionalProperties": false,
            "required": ["scope", "resetAt"],
            "properties": {
              "scope": { "type": "string", "minLength": 1, "maxLength": 255 },
              "limit": { "type": "integer", "minimum": 0 },
              "remaining": { "type": "integer", "minimum": 0 },
              "resetAt": { "$ref": "#/components/schemas/UtcTimestamp" }
            }
          }
        }
      }
    }
  },
...
}