# services/api/src/features/iam/api/current-session.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsBoolean, IsOptional, IsUUID } from 'class-validator';

/** IAM-009: the redacted identity projection used to hydrate an authenticated client. */
export class CurrentSessionDto {
  @ApiProperty({ format: 'uuid' })
  @IsUUID()
  userId!: string;

  @ApiProperty({ format: 'uuid' })
  @IsUUID()
  organizationId!: string;

  @ApiProperty({ format: 'uuid', required: false })
  @IsOptional()
  @IsUUID()
  workspaceId?: string;

  @ApiProperty({ minimum: 1 })
  authorizationEpoch!: number;

  @ApiProperty()
  @IsBoolean()
  mfaRequired!: boolean;
}